requirements.txt
beautifulsoup4==4.8.2
python-wordpress-xmlrpc==2.3
requests==2.22.0